i386: prototypical keyboard input support
This is nowhere near done. Right now, it just prints all key codes
as ASCII characters, without accounting for held keys or locks. The
design should be slightly modified to account for this. Also, PS/2
extended scancodes (those that are introduced with 0xE0) are not
translated yet. Perhaps there's some obviously better design to
translate scancodes into keycodes that I'm missing, but this will do
for now.

I'm thinking of using /dev/kbd in the terminal to allow for the
keyboard to be read from, which would require an overhaul of the VFS
to make it work.