random: add (CS?)PRNG
This is based off of the ChaCha20 cipher, so it's marginally secure.
I say marginally because I implemented it myself. :P

Currently, the RNG's key is not randomized, and no entropy is
used whatsoever. The key is simply seeded to 0x0...0, so the
results are currently completely deterministic. This will be
corrected soon via the addition of an entropy pool and perhaps
by the addition of a real-time clock, which might allow the
current second and milliseconds spent booting to be added to
the entropy pool.