clock/i386: add clock support and x86 RTC
This allows a clock to be registered, which can be read
at any time. Note that a clock is not necessarily a timer!
Later, support should be added for unregistering clocks.
I would like to use this sort of interface more in the future
instead of the "fit arch-specific to generic include file"
interface as it lends itself well to using more than one kind
of device at a time.