libc: Simplify malloc
Back to a free list... at least this design's a little less prone to
issues. There are a couple of bugs and a hackfix, unfortunately, but
I can't exactly find them right now. They appear to require a bunch
of sequential large allocations. The nature of these bugs ranges from
a block's previous block's next block (ie. the original block) not
being equal to the original block (this was hackfixed) and the usual
weird disable-paging-and-triple-fault (maybe not malloc's fault?).