amd64: memory management subsystem!
It's a bit of a mess and needs some cleanup, but it works!
It's kinda odd because the design of the system goes against
the grain prescribed by the x86_64 crate, but it's not too
bad. I will clean it up later™, of course :P

One notable feature that's missing is lazy page frame allocation.
I've got the framework for it (a function that allocates a page w/
PageFlags::BIT_9 set like in classic Spinel), but I can't seem to
coerce x86_64 into giving me the page flags. Worst comes to worst,
I'll have to write my own page mapping abstraction