alloc: kernel memory allocator!
This isn't a very original allocator, but I was pretty mentally
exhausted when I wrote this... not a good thing for a memory allocator,
I guess. I will probably go back to give this a tune up and make it
more efficient for kernel purposes (or perhaps I'll go with the
original plan from classic Spinel and use the same allocator for
kernelspace and userspace).