general: WIP scheduler
A lot of work has been done. This probably should be broken into
smaller commits, but the scale of these changes is rather massive.
Currently, the implementation crashes upon returning from the
switch_tasks function in the new thread due to an instruction fetch
from 0x0. Something is probably wrong with setting up the new process's
stack.

There are a lot of debug prints left in for the aid of anyone who
wishes to tackle this. These should be removed before merging into
spinel-nd.